work space pattern
