+++ src/patterns/home/patterns-home.template.html
@@ -46,9 +46,9 @@
       </a>
     </div>
     <div class="usa-width-one-third sam-docs flex">
-      <a [routerLink]="['/patterns/page/sample']" class="sam-docs card">
-        <div class="title">Sample</div>
-        <span class="sam small">Sample Pattern</span>
+      <a [routerLink]="['/patterns/page/workspace']" class="sam-docs card">
+        <div class="title">Workspace</div>
+        <span class="sam small">Workspace Pattern</span>
       </a>
     </div>
   </div>
