--- src/patterns/pages/sample/datasource.ts
@@ -1,2 +1,0 @@
-/* tslint:disable */
-export const SampleData = {"_embedded":{"results":[{"postAward":{"audit":{"questions":[{"questionCode":"subpartD","isSelected":"true"},{"questionCode":"subpartF","isSelected":"true"},{"questionCode":"subpartB","isSelected":"true"},{"questionCode":"subpartC","isSelected":"true"},{"questionCode":"subpartE","isSelected":"true"}]},"accomplishments":[{"year":"2008","description":"Their have been over 300 fellows selected since 1992. The Pickering Fellowship attracts outstanding students who represent all ethnic and social backgrounds, and who have an interest in pursuing a Foreign Service Career in the U.S. Department of State. This program develops a source of trained men and women who are dedicated to representing America's interests abroad."},{"year":"2009","description":"No current data available."},{"year":"2010","description":"No current data available."}]},"website":"http://www.careers.state.gov/students/programs.html#TRP","financial":{"additionalInfo":"Please see obligations...","obligations":[{"assistanceType":{"code":"0090001","value":"Salaries and Expenses"},"isRecoveryAct":false,"values":[{"actual":6042544,"year":2008},{"year":2010,"estimate":6555091},{"year":2009,"estimate":6242654}],"additionalInfo":null}]},"suggestion":{"input":["thomas r. pickering foreign affairs fellowship program","19.013"],"context":{"isActive":true}},"usage":{"discretionaryFund":null,"loanTerms":null,"restrictions":"Funding is to used to support program administration as well as coordinating program logistics. Funding supports activities and financial obligations such as tuition costs for undergraduate and graduate study; room and board; student travel; books, fees and costs associated with student summer internship programs including stipends, travel, passport fees, inoculation and visa fees. Additionally, a designated amount of funds may be used for administrative costs associated with the program administration.","rules":null},"_rScore":0,"publishDate":"2018-02-22T11:35:42-05:00","_type":"assistanceListing","eligibility":{"beneficiary":{"types":[{"code":"33","value":"Specialized group (e.g. health professionals, students, veterans)"},{"code":"37","value":"Quasi-public nonprofit organization"},{"code":"63","value":"U.S. Citizen"},{"code":"21","value":"Other public institution/organization"},{"code":"36","value":"Private nonprofit institution/organization"},{"code":"20","value":"Public nonprofit institution/organization"},{"code":"42","value":"Education Professional"},{"code":"32","value":"Minority group"}],"additionalInfo":null},"documentation":"Applications may be submitted by public and private non-profit organizations. Please refer to Grants.gov for further eligibility criteria.","applicant":{"types":[{"code":"37","value":"Quasi-public nonprofit institution/organization"},{"code":"36","value":"Private nonprofit institution/organization (includes institutions of higher education and hospitals)"},{"code":"30","value":"Non-Government - General"},{"code":"21","value":"Other public institution/organization"},{"code":"32","value":"Minority group"},{"code":"20","value":"Public nonprofit institution/organization (includes institutions of higher education and hospitals)"}],"additionalInfo":null,"assistanceUsageTypes":null}},"assistanceTypes":[{"hierarchy":[{"elementId":"0003","code":"B","level":1,"value":"PROJECT GRANTS"},{"elementId":"0003001","code":"B","level":2,"value":"Cooperative Agreements"}]},{"hierarchy":[{"elementId":"0003","code":"B","level":1,"value":"PROJECT GRANTS"},{"elementId":"0003002","code":"B","level":2,"value":"Cooperative Agreements (Discretionary Grants)"}]},{"hierarchy":[{"elementId":"0003","code":"B","level":1,"value":"PROJECT GRANTS"},{"elementId":"0003009","code":"B","level":2,"value":"Project Grants (Fellowships)"}]}],"indexedDate":"2018-02-27T14:30:33-05:00","title":"Thomas R. Pickering Foreign Affairs Fellowship Program","isActive":true,"objective":"As authorized by authorizations stated above-the The Pickering Program has an undergraduate and graduate component.  Both programs have a common purpose - to attract outstanding students who represent all ethnic and social backgrounds, who have an interest in pursuing a Foreign Service career in the U.S. Department of State. These programs develop a source of trained men and women, from academic fields representing the skill needs of the Department who are dedicated to representing America's interests abroad. ","alternativeNames":["Pickering Fellowship Program"],"isFunded":false,"historicalIndex":[{"organizationId":"100005153","actionType":"publish","createdDate":"2018-02-22T11:35:42.091-05:00","isManual":false,"index":2,"historicalIndexId":"89a54c09279a498d99464fd2d9357244","body":"Thomas R. Pickering Foreign Affairs Fellowship Program","fiscalYear":2018,"statusCode":""},{"organizationId":"100005153","actionType":"publish","createdDate":"2009-07-09T18:03:13-04:00","isManual":false,"index":1,"historicalIndexId":"72ff7483fe8809c1c9f98a43f4264554","body":"Thomas R. Pickering Foreign Affairs Fellowship Program","fiscalYear":2009,"statusCode":null}],"application":{"selectionCriteria":"Quality of program idea; Program planning ability; Ability to achieve objectives; Multiplier effect/impact i.e. can the program support an increased load; Support diversity; International capacity; institution's record/ability; Cost-effectiveness/savings. Please refer to Grants.gov for further information."},"isLatest":true,"modifiedDate":"2018-02-22T11:35:42-05:00","programNumber":"19.013","organizationHierarchy":[{"organizationId":"100004222","level":1,"name":"HEALTH AND HUMAN SERVICES, DEPARTMENT OF"},{"organizationId":"100005153","level":2,"name":"OFFICE OF ASSISTANT SECRETARY FOR PERSONNEL"}],"_id":"b57c8823287e45dc8619937454ad0b6c","contacts":[{"address":{"zip":"20522","country":"US","streetAddress":"Dept of State SA-1,Student Programs, Rm H-518, 2401 E Street NW\r\n","city":"Washington","state":"DC"},"phone":"1111111111","name":"Ebony K. Smith,","title":"HR Specialist - Pickering Fellowship Program"}]},{"postAward":{"audit":{"questions":[{"questionCode":"subpartF","isSelected":"false"},{"questionCode":"subpartE","isSelected":"false"},{"questionCode":"subpartD","isSelected":"true"},{"questionCode":"subpartC","isSelected":"true"},{"questionCode":"subpartB","isSelected":"true"}]},"accomplishments":null},"financial":{"additionalInfo":"Awards ranged from $980,000 to $3,800,000.  Awards we based on per-capita population and the burden of disease of diabetes and heart disease and stroke in each individual state. ","obligations":[{"assistanceType":{"code":"0003003","value":"Project Grants"},"isRecoveryAct":false,"values":[{"year":2017,"estimate":0},{"year":2018,"estimate":120769531},{"actual":0,"year":2016}],"additionalInfo":null}]},"suggestion":{"input":["improving the health of americans through prevention and management of diabetes and heart disease and stroke","93.426"],"context":{"isActive":true}},"_rScore":0,"publishDate":"2017-11-26T05:00:03-05:00","preApplication":{"coordination":{"reports":[{"reportCode":"statement","isSelected":false},{"reportCode":"assessment","isSelected":false},{"reportCode":"ExecutiveOrder12372","isSelected":true},{"reportCode":"otherRequired","isSelected":true}],"description":null}},"_type":"assistanceListing","eligibility":{"beneficiary":{"types":[{"code":"14","value":"State"}],"additionalInfo":"Beneficiaries of this program include: State, Local, Individual/Family, Minority Group, Anyone/General Public, Black American, American Indian, Spanish Origin, Oriental, Other Nonwhite, Women, Handicapped, Physically Afflicted, Senior Citizen, Rural"},"documentation":null,"applicant":{"types":[{"code":"14","value":"State (includes District of Columbia, public institutions of higher education and hospitals)"}],"additionalInfo":"State Governments or their Bona Fide Agents (includes the District of Columbia) are eligible","assistanceUsageTypes":null}},"assistanceTypes":[{"hierarchy":[{"elementId":"0003","code":"B","level":1,"value":"PROJECT GRANTS"},{"elementId":"0003001","code":"B","level":2,"value":"Cooperative Agreements"}]}],"indexedDate":"2018-02-27T14:30:33-05:00","relatedPrograms":[{"programNumber":"93.757","id":"258d597fbc3dba2418fb057eb07e9332","title":"State and Local Public Health Actions to Prevent Obesity, Diabetes, Heart Disease and Stroke (PPHF)"}],"title":"Improving the Health of Americans through Prevention and Management of Diabetes and Heart Disease and Stroke","isActive":true,"objective":"The purpose of this program is to implement and evaluate evidence-based strategies to address the challenges and systemic barriers that contribute to prevention and management of cardiovascular disease and diabetes in high-burden populations.  The Selection of high burden populations will be based on a robust analysis of diabetes and cardiovascular health burden across geographic areas and population subgroups. Population subgroups may be defined by factors such as race or ethnicity, gender, age [e.g. youth and older adults], education, income, disability, or sexual orientation, among others.   ","alternativeNames":["The purpose of this program is to implement and evaluate evidence-based strategies to address the challenges and systemic barriers that contribute to prevention and management of cardiovascular disease and diabetes in high-burden populations.  The Selection of high burden populations will be based on a robust analysis of diabetes and cardiovascular health burden across geographic areas and population subgroups. Population subgroups may be defined by factors such as race or ethnicity, gender, age [e.g. youth and older adults], education, income, disability, or sexual orientation, among others.   "],"isFunded":true,"historicalIndex":[{"organizationId":"100007448","actionType":"publish","createdDate":"2017-11-26T05:00:03-05:00","isManual":false,"index":1,"historicalIndexId":"5981767343fec03abb149ba33b88d733","body":"Improving the Health of Americans through Prevention and Management of Diabetes and Heart Disease and Stroke","fiscalYear":2017,"statusCode":null}],"application":{"selectionCriteria":" Review and Selection Process: Applications will be reviewed in three phases.\r\n\r\nPhase I Review: All applications will be reviewed initially for completeness by CDC OGS staff and will be reviewed jointly for eligibility by the CDC National Center for Chronic Disease Prevention and Health Promotion and CDC OGS.  Incomplete applications that do not meet the eligibility criteria will not advance to Phase II review.\r\n\r\nPhase II Review: This funding opportunity is comprised of two components – Component 1 and Component 2.  Applications submitted under Component 1 are non-competitive and will support the implementation and evaluation of a core set of evidence-based strategies defined to reduce risk, complications, and/or barriers to prevent and control diabetes and cardiovascular disease among high-burden populations in all states and the District of Columbia.  Applications submitted under Component 2 will be competitive and will support the implementation and evaluation of work on a small set of innovative strategies designed to reduce risks, complications, and/or barriers to prevent and control diabetes and prevention of heart disease and stroke.  Component 2 applications will undergo an objective review panel process to evaluate complete and responsiveness according to the criteria section of the funding opportunity.\r\n\r\nPhase III Review:  Applications submitted under Component 1 will be funded in accordance with the number of strategies selected by applicants from each of the diabetes and heart disease so that investments and outcomes may be clearly identified by funding line.  Applications submitted under Component 2 will be funded in order by score and rank determined by the objective review panel.  The Selecting official shall rely on the rank order established by the objective review panel as the primary factor in making awards.  \r\n\r\n"},"isLatest":true,"modifiedDate":"2017-11-26T05:00:03-05:00","programNumber":"93.426","organizationHierarchy":[{"organizationId":"100004222","level":1,"name":"HEALTH AND HUMAN SERVICES, DEPARTMENT OF"},{"organizationId":"100007448","level":2,"name":"CENTERS FOR DISEASE CONTROL AND PREVENTION"}],"_id":"07dbe7d1b296db06bd592215dc0a572e","contacts":[{"address":{"zip":"30341","country":"US","streetAddress":"4770 Buford Highway, MS F75","city":"Atlanta","state":"GA"},"phone":"7704881431","name":"Robert Hancock","title":null}]},{"postAward":{"audit":{"questions":[{"questionCode":"subpartE","isSelected":"false"},{"questionCode":"subpartB","isSelected":"true"},{"questionCode":"subpartF","isSelected":"false"},{"questionCode":"subpartC","isSelected":"true"},{"questionCode":"subpartD","isSelected":"true"}]},"accomplishments":null},"website":"http://www.cdc.gov","financial":{"additionalInfo":"$2,745,067 to $3,000,000.","obligations":[{"assistanceType":{"code":"0003001","value":"Cooperative Agreements"},"isRecoveryAct":false,"values":[{"year":2018,"estimate":2745067},{"year":2017,"estimate":2745067},{"actual":2745067,"year":2016}],"additionalInfo":null}]},"suggestion":{"input":["cdc undergraduate public health scholars program (cups): a public health experience to expose undergraduates interested in minority health to public health and the public health professions","93.456"],"context":{"isActive":true}},"_rScore":0,"publishDate":"2017-11-25T05:00:03-05:00","_type":"assistanceListing","eligibility":{"beneficiary":{"types":[{"code":"38","value":"Other private institution/organization"},{"code":"36","value":"Private nonprofit institution/organization"},{"code":"32","value":"Minority group"},{"code":"21","value":"Other public institution/organization"},{"code":"20","value":"Public nonprofit institution/organization"},{"code":"18","value":"Sponsored organization"}],"additionalInfo":"OMHHE will collaborate with educational institutions, including those serving minority populations, to increase the knowledge, diversity, and skills of students in public health through internship and fellowship programs."},"documentation":null,"applicant":{"types":[{"code":"36","value":"Private nonprofit institution/organization (includes institutions of higher education and hospitals)"},{"code":"30","value":"Non-Government - General"},{"code":"21","value":"Other public institution/organization"},{"code":"38","value":"Other private institutions/organizations"}],"additionalInfo":"Funds are to be used for recruitment, orientation, placement, mentorship and follow-up tracking of undergraduate and graduate students. ","assistanceUsageTypes":null}},"assistanceTypes":[{"hierarchy":[{"elementId":"0003","code":"B","level":1,"value":"PROJECT GRANTS"},{"elementId":"0003001","code":"B","level":2,"value":"Cooperative Agreements"}]}],"indexedDate":"2018-02-27T14:30:33-05:00","title":"CDC Undergraduate Public Health Scholars Program (CUPS): A Public Health Experience to Expose Undergraduates Interested in Minority Health to Public Health and the Public Health Professions","isActive":true,"objective":"The purpose of this funding opportunity is to implement a national summer training program to introduce undergraduate and graduate students, including but not limited to those from under-represented and underserved racial and ethnic minority populations, to public health and biomedical sciences.","isFunded":true,"historicalIndex":[{"organizationId":"100007448","actionType":"publish","createdDate":"2017-02-02T05:00:09-05:00","isManual":false,"index":1,"historicalIndexId":"fa291ad8d609d3d7f818e4d766819a41","body":"CDC Undergraduate Public Health Scholars Program (CUPS): A Public Health Experience to Expose Undergraduates Interested in Minority Health to Public Health and the Public Health Professions","fiscalYear":2017,"statusCode":null}],"application":{"selectionCriteria":"Complete review criteria is published in the Funding Opportunity Announcement (FOA) in www.grants.gov. Listed are the criteria used to review applications: (1) approach; (2) evaluation and performance measurement; and (3) applicant's organizational capacity to implement the approach."},"isLatest":true,"modifiedDate":"2017-11-25T05:00:03-05:00","programNumber":"93.456","organizationHierarchy":[{"organizationId":"100004222","level":1,"name":"HEALTH AND HUMAN SERVICES, DEPARTMENT OF"},{"organizationId":"100007448","level":2,"name":"CENTERS FOR DISEASE CONTROL AND PREVENTION"}],"_id":"273825f215c061ee899be7448ecf016c","contacts":[{"address":{"zip":"30013","country":"US","streetAddress":"2900 Woodcock Boulevard ","city":"Atlanta","state":"GA"},"phone":"7704881387","name":"Glynnis Taylor","title":"Public Health Analyst"}]},{"postAward":{"audit":{"questions":[{"questionCode":"subpartF","isSelected":"false"},{"questionCode":"subpartD","isSelected":"true"},{"questionCode":"subpartC","isSelected":"true"},{"questionCode":"subpartB","isSelected":"true"},{"questionCode":"subpartE","isSelected":"false"}]},"accomplishments":null},"website":"http://www.grants.gov","financial":{"additionalInfo":"0 to 75000","obligations":[{"assistanceType":{"code":"0090001","value":"Salaries and Expenses"},"isRecoveryAct":false,"values":[{"year":2018,"estimate":75000},{"year":2017,"estimate":75000},{"actual":0,"year":2016}],"additionalInfo":null}]},"suggestion":{"input":["national ocean service intern program","11.480"],"context":{"isActive":true}},"_rScore":0,"publishDate":"2017-11-24T05:00:02-05:00","_type":"assistanceListing","eligibility":{"beneficiary":{"types":[{"code":"45","value":"Scientist/Researchers"},{"code":"43","value":"Student/Trainee"},{"code":"39","value":"Anyone/general public"}],"additionalInfo":"Benefits to the Nation are increased opportunities for recent college graduates to increase public awareness and gain exposure to government programs that affect the Nation's coastal areas and resources. This program will also promote interest in civil service as a career. Benefits also include a transfer of the techniques, technologies, and methods used by NOS to the next generation of engineers, scientists, and managers in the environmental arena. Interns who are selected by the program will receive increased training and skill enhancement as identified by the particular Internship opportunity in which they participate."},"documentation":"Applicants are required to satisfy all Department of Commerce (DOC) and National Oceanic and Atmospheric Administration (NOAA) standards and regulations, including routine and special terms and conditions, for financial assistance programs application and conduct. Costs will be determined in accordance with OMB Circular Nos. A-21 for educational institutions and A-22 for nonprofit organizations.","applicant":{"types":[{"code":"20","value":"Public nonprofit institution/organization (includes institutions of higher education and hospitals)"}],"additionalInfo":"Primary awards will be available to nonprofit organizations. Internships are restricted to recent college graduates and have a maximum cumulative duration of 24 months for any individual.","assistanceUsageTypes":null}},"assistanceTypes":[{"hierarchy":[{"elementId":"0003","code":"B","level":1,"value":"PROJECT GRANTS"},{"elementId":"0003003","code":"B","level":2,"value":"Project Grants"}]}],"indexedDate":"2018-02-27T14:30:33-05:00","title":"National Ocean Service Intern Program","isActive":true,"objective":"To establish an Intern Program which will provide unique opportunities for cooperative study, research, and development that would be of major benefit in advancing the number and diversity of skilled engineers, scientists, and managers in the environmental arena who are familiar with the techniques and technologies used by the National Ocean Service (NOS).","alternativeNames":["(NOS Intern Program)"],"isFunded":true,"historicalIndex":[{"organizationId":"100099213","actionType":"unarchive","createdDate":"2017-10-17T13:18:52-04:00","isManual":false,"index":3,"historicalIndexId":"4cd7d3c1f18bd587fb6de933b15fb3bf","body":"Reinstated","fiscalYear":2017,"statusCode":null},{"organizationId":"100099213","actionType":"archive","createdDate":"2009-07-07T15:38:42-04:00","isManual":false,"index":2,"historicalIndexId":"03bbad05ad7d4b320da51f01d88b4a82","body":"Archived","fiscalYear":2009,"statusCode":null},{"organizationId":"100099213","actionType":"publish","createdDate":"2004-06-21T00:00:00-04:00","isManual":true,"index":1,"historicalIndexId":"f72b2c8c77542cbc6a9d24eb2d6ca72d","body":"National Ocean Service Intern Program","fiscalYear":1999,"statusCode":"B"}],"isLatest":true,"modifiedDate":"2017-11-24T05:00:02-05:00","programNumber":"11.480","organizationHierarchy":[{"organizationId":"100035122","level":1,"name":"COMMERCE, DEPARTMENT OF"},{"organizationId":"100099213","level":2,"name":"NATIONAL OCEANIC AND ATMOSPHERIC ADMINISTRATION"}],"_id":"141a28376ab625713b06ff8c545c6944","contacts":[{"address":{"zip":"20910","country":"US","streetAddress":"1305 East West Highway\r\nSSMC 4 Room 13267","city":"Silver Spring","state":"MD"},"phone":"2405330955","name":"Kadija Baffoe-Harding","title":"NOS GMAC"}]},{"postAward":{"audit":{"questions":[{"questionCode":"subpartB","isSelected":"true"},{"questionCode":"subpartF","isSelected":"true"},{"questionCode":"subpartE","isSelected":"true"},{"questionCode":"subpartD","isSelected":"true"},{"questionCode":"subpartC","isSelected":"true"}]},"accomplishments":null},"website":"http://www.cdc.gov","financial":{"additionalInfo":"Subject to the availability of funds; Fiscal year 2018-2022; Approximate total fiscal year funding is $6,500,000; Approximate Total Project Period Funding: $31,250,000 over five years; Approximate number of awards is 25 total as follows:\r\n\r\n• Category A: up to 8 awards\r\n• Category B: up to 8 awards\r\n• Category C: up to 9 awards\r\n\r\nEstimated Award Floor is $20,000 with ceiling of individual award using calculation below:\r\n\r\n=\r\n[Base program office's funding (up to $50,000) + Sum of other CDC programs (various annually)] / # of awardees","obligations":[{"assistanceType":{"code":"0003001","value":"Cooperative Agreements"},"isRecoveryAct":false,"values":[{"year":2018,"estimate":6500000},{"flag":"ena","year":2016},{"flag":"ena","year":2017}],"additionalInfo":null}]},"suggestion":{"input":["tribal public health capacity building and quality improvement umbrella cooperative agreement","93.772"],"context":{"isActive":true}},"_rScore":0,"publishDate":"2017-11-23T05:00:03-05:00","_type":"assistanceListing","eligibility":{"beneficiary":{"types":[{"code":"58","value":"American Indian"},{"code":"40","value":"Native American Organizations"},{"code":"22","value":"Federally Recognized Indian Tribal Governments"}],"additionalInfo":"Eligible applicants should be able to demonstrate tribal affiliation including (1) a record of effectively working with American Indian and Alaska Native populations (2) an ability to methodically and efficiently reach tribal members in American Indian and Alaska Native communities and (3), if recipient is a tribally owned and operated organization, it should provide a letter of approval from Tribal council for the proposed program. "},"documentation":"For non-profits, proof of non-profit IRS status should be submitted with application documents","applicant":{"types":[{"code":"22","value":"Federally Recognized lndian Tribal Governments"},{"code":"40","value":"Native American Organizations (includes lndian groups, cooperatives, corporations, partnerships, associations)"}],"additionalInfo":"\r\n\r\nAny regional native-serving organizations whose board is comprised of federally recognized Indian Tribes and elected/appointed Tribal leaders, including. \r\n\r\n1.Inter-Tribal consortia, \r\n2.Urban Indian organizations, \r\n3.Regional Indian health boards, and \r\n4.Tribally owned and operated health systems. \r\n\r\n","assistanceUsageTypes":null}},"assistanceTypes":[{"hierarchy":[{"elementId":"0003","code":"B","level":1,"value":"PROJECT GRANTS"},{"elementId":"0003001","code":"B","level":2,"value":"Cooperative Agreements"}]}],"indexedDate":"2018-02-27T14:30:33-05:00","relatedPrograms":[{"programNumber":"93.098","id":"9713b3ac8e2651310d761f4dab3c825e","title":"Tribal Public Health Capacity Building and Quality Improvement"},{"programNumber":"93.331","id":"fec1ab7e76a9eb5be67eeba365c1bef2","title":"Partnerships to Improve Community Health"}],"title":"Tribal Public Health Capacity Building and Quality Improvement Umbrella Cooperative Agreement","isActive":true,"objective":"To strengthen and improve the infrastructure and performance of tribal public health systems  through the optimization of  workforce, data surveillance and information systems, development/adaptation of evidence-based/informed interventions, and strategic partnerships. Primary objectives include (1) advanced capacity to identify, respond to and remediate public health threats; (2) decrease burden of disease and (3) improve Tribal public health policies to protect population health.","alternativeNames":["Tribal Umbrella"],"isFunded":true,"historicalIndex":[{"organizationId":"100007448","actionType":"publish","createdDate":"2017-08-02T05:00:20-04:00","isManual":false,"index":1,"historicalIndexId":"a52d48ca37247d0c31a6a1619527d04d","body":"Tribal Public Health Capacity Building and Quality Improvement Umbrella Cooperative Agreement","fiscalYear":2017,"statusCode":null}],"application":{"selectionCriteria":"As specified in the funding opportunity announcement OT18-1803, Section II E, application review and Selection Process. An objective review panel with experience working with tribes or tribal organizations will evaluate complete applications. Reviewers will be primarily selected from a pool of CDC subject matter experts with demonstrated experience working with tribes or in tribal settings to establish cultural competence. The specific objective reviewer scores will be used to produce a rank order of proposals. However, proposals may be funded out of rank order based upon the following criteria: health conditions and geographical diversity of populations that the applicants will directly serve. "},"isLatest":true,"modifiedDate":"2017-11-23T05:00:03-05:00","programNumber":"93.772","organizationHierarchy":[{"organizationId":"100004222","level":1,"name":"HEALTH AND HUMAN SERVICES, DEPARTMENT OF"},{"organizationId":"100007448","level":2,"name":"CENTERS FOR DISEASE CONTROL AND PREVENTION"}],"_id":"ed2e18fb78b771a73f49eb99d0db2971","contacts":[{"address":{"zip":"30345","country":"US","streetAddress":"1825 Century Blvd NE","city":"Atlanta","state":"GA"},"phone":"404-498-2208","name":"Carmen (Skip) Clelland","title":"Associate Director, Tribal Support"}]},{"postAward":{"audit":{"questions":[{"questionCode":"subpartB","isSelected":"true"},{"questionCode":"subpartC","isSelected":"true"},{"questionCode":"subpartD","isSelected":"true"},{"questionCode":"subpartE","isSelected":"true"},{"questionCode":"subpartF","isSelected":"true"}]},"accomplishments":null},"financial":{"additionalInfo":"Average award $810,000","obligations":[{"assistanceType":{"code":"0090001","value":"Salaries and Expenses"},"isRecoveryAct":false,"values":[{"year":2018,"estimate":18000000},{"flag":"ena","year":2016,"explanation":"new proram"},{"flag":"ena","year":2017}],"additionalInfo":null}]},"suggestion":{"input":["flexible funding model - infrastructure development and maintenance for state manufactured food regulatory programs","93.367"],"context":{"isActive":true}},"usage":{"discretionaryFund":null,"loanTerms":null,"restrictions":"No vehicle purchase\r\nNo new construction\r\n","rules":null},"_rScore":0,"publishDate":"2017-11-18T05:00:02-05:00","_type":"assistanceListing","eligibility":{"beneficiary":{"types":[{"code":"14","value":"State"}],"additionalInfo":"N/A"},"documentation":"Competitive application submission to appropriate RFA's","applicant":{"types":[{"code":"24","value":"U.S. Territories and possessions"},{"code":"22","value":"Federally Recognized lndian Tribal Governments"},{"code":"13","value":"State"}],"additionalInfo":"N/A","assistanceUsageTypes":null}},"assistanceTypes":[{"hierarchy":[{"elementId":"0003","code":"B","level":1,"value":"PROJECT GRANTS"},{"elementId":"0003001","code":"B","level":2,"value":"Cooperative Agreements"}]}],"indexedDate":"2018-02-27T14:30:33-05:00","relatedPrograms":[{"programNumber":"93.103","id":"e87c436bbb5a61c39042d6a5d1620bdc","title":"Food and Drug Administration Research"}],"title":"Flexible Funding Model - Infrastructure Development and Maintenance for State Manufactured Food Regulatory Programs","isActive":true,"objective":"Advance efforts for a nationally integrated food safety system (IFSS) by supporting Manufactured Food Regulatory Program Standards (MFRPS), Rapid Response Teams (RRT) and Food Protection Task Force (FPTF) programs, as well as special projects. ","alternativeNames":["Flexible Funding Model - Infrastructure Development and Maintenance for State Manufactured Food Regulatory Programs"],"isFunded":true,"historicalIndex":[{"organizationId":"100008023","actionType":"publish","createdDate":"2017-11-18T05:00:03-05:00","isManual":false,"index":1,"historicalIndexId":"89170c18183ace877b5b2129acb23073","body":"Flexible Funding Model - Infrastructure Development and Maintenance for State Manufactured Food Regulatory Programs","fiscalYear":2017,"statusCode":null}],"application":{"selectionCriteria":"Listed in grant announcement upon posting"},"isLatest":true,"modifiedDate":"2017-11-18T05:00:02-05:00","programNumber":"93.367","organizationHierarchy":[{"organizationId":"100004222","level":1,"name":"HEALTH AND HUMAN SERVICES, DEPARTMENT OF"},{"organizationId":"100008023","level":2,"name":"FOOD AND DRUG ADMINISTRATION"}],"_id":"a7ff7097f3debfa2384ee0b90c63aea3","contacts":[{"address":{"zip":"20857","country":"US","streetAddress":"5630 Fishers Lane\r\n","city":"Rockville","state":"MD"},"phone":"2404027596","name":"Daniel Lukash","title":null}]},{"postAward":{"audit":{"questions":[{"questionCode":"subpartF","isSelected":"true"},{"questionCode":"subpartD","isSelected":"true"},{"questionCode":"subpartC","isSelected":"true"},{"questionCode":"subpartE","isSelected":"false"},{"questionCode":"subpartB","isSelected":"true"}]},"accomplishments":null},"website":"http://www.state.gov/g/oes/","financial":{"additionalInfo":"Grants vary depending on program objectives and availability of funds.","obligations":[{"assistanceType":{"code":"0003003","value":"Project Grants"},"isRecoveryAct":false,"values":[{"flag":"ena","year":2018},{"actual":560206000,"year":2016},{"year":2017,"estimate":7100000}],"additionalInfo":"n/a"}]},"suggestion":{"input":["environmental and scientific partnerships and programs","19.017"],"context":{"isActive":true}},"usage":{"discretionaryFund":null,"loanTerms":null,"restrictions":"Various restrictions may apply, depending on the source of funding.  Please refer to the individual program announcement in grants.gov for further information. For profit-organizations under exceptional circumstances may be available and are subject of prior approval.","rules":null},"_rScore":0,"publishDate":"2017-11-17T05:00:03-05:00","_type":"assistanceListing","eligibility":{"beneficiary":{"types":[{"code":"45","value":"Scientist/Researchers"},{"code":"21","value":"Other public institution/organization"},{"code":"20","value":"Public nonprofit institution/organization"},{"code":"38","value":"Other private institution/organization"},{"code":"36","value":"Private nonprofit institution/organization"}],"additionalInfo":"n/a"},"documentation":null,"applicant":{"types":[{"code":"21","value":"Other public institution/organization"},{"code":"30","value":"Non-Government - General"},{"code":"20","value":"Public nonprofit institution/organization (includes institutions of higher education and hospitals)"},{"code":"36","value":"Private nonprofit institution/organization (includes institutions of higher education and hospitals)"},{"code":"38","value":"Other private institutions/organizations"},{"code":"35","value":"Profit organization"}],"additionalInfo":"The use of assistance is to support a wide range of partnerships responding to diverse  international environmental initiatives. ","assistanceUsageTypes":null}},"assistanceTypes":[{"hierarchy":[{"elementId":"0003","code":"B","level":1,"value":"PROJECT GRANTS"},{"elementId":"0003003","code":"B","level":2,"value":"Project Grants"}]},{"hierarchy":[{"elementId":"0003","code":"B","level":1,"value":"PROJECT GRANTS"},{"elementId":"0003001","code":"B","level":2,"value":"Cooperative Agreements"}]}],"indexedDate":"2018-02-27T14:30:33-05:00","title":"Environmental and Scientific Partnerships and Programs","isActive":true,"objective":"To support sustainable growth across the globe, the Bureau of Oceans, Environment, and Science (OES) builds international partnerships to address environmental, scientific, and health concerns.  Key areas of funding include but are not limited to, programs that: address climate change and its impact, and reduce global greenhouse gas emissions; conserve natural resources and reduce international threats to biodiversity;  improve access to safe drinking water;  promote a level playing field with free trade partners; protect fisheries and oceans; reduce the threat from mercury and other pollutants; and foster international scientific collaboration.","isFunded":true,"historicalIndex":[{"organizationId":"100120039","actionType":"publish","createdDate":"2009-10-14T10:32:46-04:00","isManual":false,"index":1,"historicalIndexId":"935b36a6b335f4b24c163a95ead5f108","body":"Environmental and Scientific Partnerships and Programs","fiscalYear":2009,"statusCode":null}],"application":{"selectionCriteria":"Proposals should be responsive to the objectives and requirements set forth in each Program Announcement or Request for application posted on www.grants.gov"},"isLatest":true,"modifiedDate":"2017-11-17T05:00:03-05:00","programNumber":"19.017","_id":"b9d213e77ad5cf90666bc085becd90da","contacts":[{"address":{"zip":"22520","country":"US","streetAddress":"Central Programming Unit, Executive Office, Bureau of Oceans, International Environmental and Scientific Affairs, Room 10076  SA-09, Department of State, 2025 E Street NW ","city":"Washington","state":"DC"},"phone":"(202) 472-8171","name":"Maria Urbina","title":"Program Manager"}]},{"projects":[{"year":"2017","description":"Fish, Coasts, Ocean Exploration, Weather, Research"},{"year":"2018","description":"Fish, Coasts, Ocean Exploration, Weather, Research"},{"year":"2016","description":"Unknown Fish, Coasts, Ocean Exploration, Weather, Research"}],"financial":{"additionalInfo":"$5,000 -$4.2 million","obligations":[{"assistanceType":{"code":"0003003","value":"Project Grants"},"isRecoveryAct":false,"values":[{"flag":"nsi","year":2016},{"flag":"ena","year":2018},{"year":2017,"estimate":13313601}],"additionalInfo":"The BAA is simply a vehicle to enable receipt of applications for innovative projects.  Projects selected for funding carry the CFDA number of the program to which the work outlined in the project is most closely aligned.  "}]},"usage":{"discretionaryFund":"Funding is discretionary; however, it is based on availability.","loanTerms":null,"restrictions":"Funding is specifically for work not addressed through NOAA's competitive discretionary programs.  Funding is based on availability. ","rules":null},"publishDate":"2017-11-16T05:00:03-05:00","eligibility":{"beneficiary":{"types":[{"code":"44","value":"Graduate Student"},{"code":"59","value":"Spanish Origin"},{"code":"45","value":"Scientist/Researchers"},{"code":"57","value":"Black"},{"code":"42","value":"Education Professional"},{"code":"14","value":"State"},{"code":"18","value":"Sponsored organization"},{"code":"40","value":"Native American Organizations"},{"code":"23","value":"U.S. Territories"},{"code":"60","value":"Asian"},{"code":"58","value":"American Indian"},{"code":"31","value":"Individual/Family"},{"code":"39","value":"Anyone/general public"},{"code":"34","value":"Small business"},{"code":"33","value":"Specialized group (e.g. health professionals, students, veterans)"},{"code":"32","value":"Minority group"},{"code":"61","value":"Other Non-White"},{"code":"63","value":"U.S. Citizen"},{"code":"15","value":"Local"},{"code":"78","value":"School"},{"code":"43","value":"Student/Trainee"},{"code":"20","value":"Public nonprofit institution/organization"},{"code":"38","value":"Other private institution/organization"},{"code":"22","value":"Federally Recognized Indian Tribal Governments"},{"code":"21","value":"Other public institution/organization"},{"code":"35","value":"Profit organization"},{"code":"37","value":"Quasi-public nonprofit organization"},{"code":"36","value":"Private nonprofit institution/organization"}],"additionalInfo":"None"},"documentation":"Registration in the System for Award Management (SAM) and Grants. gov is required.  Also, applicants must submit all required Office of Management and Budget and Department of Commerce approved forms, and all required supporting documentation that thereby comprise a complete funding application package.  Additionally, applicants must adhere to all requirements contained in the Broad Agency Announcement.","applicant":{"types":[{"code":"21","value":"Other public institution/organization"},{"code":"18","value":"Sponsored organization"},{"code":"35","value":"Profit organization"},{"code":"33","value":"Specialized group (e.g. health professionals, students, veterans)"},{"code":"31","value":"Individual/Family"},{"code":"20","value":"Public nonprofit institution/organization (includes institutions of higher education and hospitals)"},{"code":"24","value":"U.S. Territories and possessions"},{"code":"34","value":"Small business (less than 500 employees)"},{"code":"15","value":"Local (includes State-designated lndian Tribes, excludes institutions of higher education and hospitals"},{"code":"14","value":"State (includes District of Columbia, public institutions of higher education and hospitals)"},{"code":"32","value":"Minority group"},{"code":"30","value":"Non-Government - General"},{"code":"22","value":"Federally Recognized lndian Tribal Governments"},{"code":"23","value":"U.S. Territories and possessions (includes institutions of higher education and hospitals)"},{"code":"37","value":"Quasi-public nonprofit institution/organization"},{"code":"36","value":"Private nonprofit institution/organization (includes institutions of higher education and hospitals)"},{"code":"13","value":"State"},{"code":"38","value":"Other private institutions/organizations"},{"code":"40","value":"Native American Organizations (includes lndian groups, cooperatives, corporations, partnerships, associations)"},{"code":"39","value":"Anyone/general public"}],"additionalInfo":"None","assistanceUsageTypes":null}},"assistanceTypes":[{"hierarchy":[{"elementId":"0003","code":"B","level":1,"value":"PROJECT GRANTS"},{"elementId":"0003001","code":"B","level":2,"value":"Cooperative Agreements"}]},{"hierarchy":[{"elementId":"0003","code":"B","level":1,"value":"PROJECT GRANTS"},{"elementId":"0003003","code":"B","level":2,"value":"Project Grants"}]}],"title":"Broad Agency Announcement","isActive":true,"objective":"This BAA is a mechanism to encourage research, education and outreach, innovative projects, or sponsorships that are not addressed through NOAA's competitive discretionary programs.  ","alternativeNames":["BAA"],"historicalIndex":[{"organizationId":"100099213","actionType":"publish","createdDate":"2015-11-06T05:03:44-05:00","isManual":false,"index":1,"historicalIndexId":"ffa47075a1a73388a3576181b6fe86d2","body":"Broad Agency Announcement","fiscalYear":2015,"statusCode":null}],"isLatest":true,"programNumber":"11.015","organizationHierarchy":[{"organizationId":"100035122","level":1,"name":"COMMERCE, DEPARTMENT OF"},{"organizationId":"100099213","level":2,"name":"NATIONAL OCEANIC AND ATMOSPHERIC ADMINISTRATION"}],"postAward":{"audit":{"questions":[{"questionCode":"subpartC","isSelected":"true"},{"questionCode":"subpartB","isSelected":"true"},{"questionCode":"subpartD","isSelected":"true"},{"questionCode":"subpartE","isSelected":"true"},{"questionCode":"subpartF","isSelected":"true"}]},"accomplishments":[{"year":"2016","description":"Projects that will be assistive in maintaining and improving the viability of marine and coastal ecosystems, delivering valuable weather, climate, and water information and services, understanding the science and consequences of climate change, and supporting global commerce and transportation. Research pertaining to fish, coasts, ocean exploration.  Projects that help to maintain and improve the viability of marine and coastal ecosystems, delivering valuable weather, climate, and water information and services, understanding the science and consequences of climate change, and supporting global commerce and transportation."},{"year":"2018","description":"Research pertaining to fish, coasts, ocean exploration.  Projects that help to maintain and improve the viability of marine and coastal ecosystems, delivering valuable weather, climate, and water information and services, understanding the science and consequences of climate change, and supporting global commerce and transportation."},{"year":"2017","description":"Projects that help to maintain and improve the viability of marine and coastal ecosystems, delivering valuable weather, climate, and water information and services, understanding the science and consequences of climate change, and supporting global commerce and transportation."}]},"website":"http://www.noaa.gov/index.html","suggestion":{"input":["broad agency announcement","11.015"],"context":{"isActive":true}},"_rScore":0,"preApplication":{"coordination":{"reports":[{"reportCode":"statement","isSelected":false},{"reportCode":"assessment","isSelected":true},{"reportCode":"ExecutiveOrder12372","isSelected":false},{"reportCode":"otherRequired","isSelected":true}],"description":"The recipient and subrecipients must comply with all environmental standards, to include those prescribed under the following statutes and Executive Orders, and shall identify to the awarding agency any impact the award may have on the environment. In some cases, funds can be withheld by the grants Officer under a special award condition requiring the recipient to submit additional environmental compliance information sufficient to enable the DOC to make an assessment on any impacts that a project may have on the environment.\r\n1. The National Environmental Policy Act of 1969 (42 U.S.C. §§ 4321 et seq.). Recipients of Federal assistance are required to identify to the awarding agency any impact an award will have on the quality of the human environment, and assist the agency in complying with the National Environmental Policy Act when the award activities remain subject to federal authority and control. Applicants for assistance may be requested to assist DOC in drafting an environmental assessment or environmental impact statement if DOC determines such documentation is required.\r\n2. The National Historic Preservation Act (NHPA) (16 U.S.C. § 470 et seq.). Recipients of Federal funding are required to identify to the awarding agency any effects the award may have on properties included on or eligible for inclusion on the National Register of Historic Places. Recipients may also be requested to assist DOC in consulting with State or Tribal Historic Preservation Officers or other applicable interested parties necessary to identify, assess and resolve adverse effects to historic properties. Until such time as the appropriate NHPA consultations and documentation is complete, funds can be withheld by the Grants Officer under a special award condition requiring the recipient to fully comply with the requirement of the NHPA.\r\n3. Executive Order 11988, May 24, 1977( “Floodplain Management”) and Executive Order 11990, May 24, 1977 (“Protection of Wetlands”). Recipients must identify proposed actions in Federally-defined floodplains and wetlands to enable the Department to make a determination as to whether there is an alternative to minimize any potential harm.\r\n4. Clean Air Act (42 U.S.C. § 7401 et seq.), Federal Water Pollution Control Act (33 U.S.C. § 1251 et seq.) (Clean Water Act), and Executive Order 11738 (“Providing for administration of the Clean Air Act and the Federal Water Pollution Control Act with respect to Federal contracts, grants or loans”). Recipients must comply with the provisions of the Clean Air Act (42 U.S.C. §§ 7401 et seq.), Clean Water Act (33 U.S.C. §§1251 et seq.), and Executive Order 11738. Recipients shall not use a facility placed by EPA on the Excluded Parties List System (EPLS) (http://www.sam.gov) in performing any award that is nonexempt under Subpart J of 2 CFR Part 1532.\r\n5. The Flood Disaster Protection Act of 1973 (42 U.S.C. § 4002 et seq.). Flood insurance, when available, is required for Federally-assisted construction or acquisition in flood-prone areas.\r\n6. The Endangered Species Act of 1973, as amended (16 U.S.C. § 1531 et seq.). Recipients must identify any impact or activities which may involve a threatened or endangered species. Federal agencies have the responsibility to ensure that no adverse effects to a protected species or habitat occur from actions under Federal assistance awards and conduct the required reviews under the Endangered Species Act, as applicable.\r\n7. The Coastal Zone Management Act, as amended (16 U.S.C. § 1451 et seq.). Funded projects must be consistent with a coastal State’s approved management program for the coastal zone.\r\n8. The Coastal Barriers Resources Act (16 U.S.C. § 3501 et seq.). Restrictions are placed on Federal funding for actions within a Coastal Barrier System.\r\n9. The Wild and Scenic Rivers Act, as amended (16 U.S.C. §§ 1271 et seq.). This Act applies to awards that may affect existing or proposed components of the National Wild and Scenic Rivers system.\r\n10. The Safe Drinking Water Act of 1974, as amended (42 U.S.C. §§ 300f-j). This Act precludes Federal assistance for any project that the EPA determines may contaminate a sole-source aquifer so as to threaten public health.\r\n11. The Resource Conservation and Recovery Act of 1976, as amended (42 U.S.C. §§ 6901 et seq.). This Act regulates the generation, transportation, treatment, and disposal of hazardous wastes. The Act also provides that recipients of Federal funds give preference in their procurement programs to the purchase of recycled products pursuant to EPA guidelines.\r\n12. The Comprehensive Environmental Response, Compensation, and Liability Act (Superfund) (42 U.S.C. § 9601 et seq.) and the Community Environmental Response Facilitation Act (41 U.S.C. § 11001 et seq.). These requirements address responsibilities of hazardous substance releases, threatened releases, and environmental cleanup. There are also reporting and community involvement requirements designed to ensure disclosure of the release or disposal of regulated substances and cleanup of hazards to state and local emergency responders.\r\n13. Executive Order 12898 (“Environmental Justice in Minority Populations and Low Income Populations”). Federal agencies are required to identify and address the disproportionately high and adverse human health or environmental effects of Federal programs, policies and activities on low income and minority populations.\r\n"}},"_type":"assistanceListing","indexedDate":"2018-02-27T14:30:33-05:00","isFunded":true,"application":{"selectionCriteria":"Research, education and outreach, innovative projects, or sponsorships that address maintaining and improving the viability of marine and coastal ecosystems, for delivering valuable weather, climate, and water information and services, for understanding the science and consequences of climate change, and for supporting the global commerce and transportation."},"modifiedDate":"2017-11-16T05:00:03-05:00","_id":"184aa9fb5b9369c49a65d11b7f57df96","contacts":[{"address":{"zip":"20910","country":"US","streetAddress":"1325 East West Highway, 9th floor, Bldg 2, Room 9328\r\n","city":"Silver Spring","state":"MD"},"phone":"3016281308","name":"Lamar Dwayne Revis","title":"Grants Officer"}]},{"projects":[{"year":"2016","description":"CDC-RFA-CK12-1203 Resident Postdoctoral Fellowship Program in Microbiology\r\n(aka ASM/CDC Program in Infectious Disease and Public Health Microbiology Postdoctoral Research Fellowship)\r\n\r\nCDC-RFA-OE15-1501 APHL-CDC Partnership: Strengthening Public Health Laboratories (EID Fellowship)"},{"year":"2018","description":"CDC-RFA-CK12-1203 Resident Postdoctoral Fellowship Program in Microbiology "},{"year":"2017","description":"CDC-RFA-CK12-1203 Resident Postdoctoral Fellowship Program in Microbiology \r\n"}],"financial":{"additionalInfo":"Program was previously funded under CDC-RFA-CK12-1203. Awards ranged from $955,013 to $1,258,661 for FY2012-FY2017.","obligations":[{"assistanceType":{"code":"0090001","value":"Salaries and Expenses"},"isRecoveryAct":false,"values":[{"actual":1258661,"year":2016},{"year":2018,"estimate":0},{"year":2017,"estimate":486126}],"additionalInfo":"Carryover of $1,079,623 unobligated balance from previous was applied to the FY17."}]},"usage":{"discretionaryFund":"This program is funded 100% by CDC discretionary budget authority.","loanTerms":null,"restrictions":"Award recipients may not use these funds for research or for directly supporting any research projects the fellows may participate in during the fellowship.","rules":null},"publishDate":"2017-11-11T05:00:03-05:00","eligibility":{"beneficiary":{"types":[{"code":"45","value":"Scientist/Researchers"},{"code":"39","value":"Anyone/general public"},{"code":"63","value":"U.S. Citizen"}],"additionalInfo":"This program contributes to developing a public health laboratory workforce for the diagnosis, prevention and control of infectious diseases in the United States and abroad. The nation's public is the ultimate recipient of benefits from this program."},"documentation":"Costs will be determined in accordance with OMB Circular No. A-87 for State and local governments. For all other nonprofit grantees, cost will be determined in accordance with HHS Regulations 45 CFR 74 Subpart Q. Title 2 of the United States Code section 1611 states that an organization described in section 501(c)(4) of the Internal Revenue Code that engages in lobbying activities is not eligible to receive Federal funds constituting an award, grant, or loan.","applicant":{"types":[{"code":"33","value":"Specialized group (e.g. health professionals, students, veterans)"},{"code":"20","value":"Public nonprofit institution/organization (includes institutions of higher education and hospitals)"}],"additionalInfo":"Eligible applicants must have experience and demonstrated success in administering training programs in public health laboratory/microbiology research or other life sciences.","assistanceUsageTypes":null}},"assistanceTypes":[{"hierarchy":[{"elementId":"0003","code":"B","level":1,"value":"PROJECT GRANTS"},{"elementId":"0003001","code":"B","level":2,"value":"Cooperative Agreements"}]}],"title":"Centers for Disease Control and Prevention - Resident Postdoctoral Program in Microbiology","isActive":true,"objective":"The purpose of this program is to conduct a resident postdoctoral fellowship program in microbiology in infectious diseases laboratories at the Centers for Disease Control and Prevention (CDC). The program offers two-year fellowships to perform research in pathogenic microorganisms. Fellows acquire specific experience in the public health aspects of microbiology, new approaches, methodologies, and knowledge in infectious disease diagnosis, prevention and control. Program goals include increasing the number of scientists trained and experienced in novel and practical applications of microbiology in infectious diseases research. In addition, goals include producing a cadre of infectious disease microbiology-research scientists in public health laboratories; increasing workforce capacity and enhancing the nation’s public health laboratory system both domestically and globally. \r\n\r\nThis program supports CDC's mission to protect America from health, safety and security threats, both foreign and in the United States. This program provides fellows research experiences that contribute to responding to and stopping new and emerging health treats, and discovering new ways to protect and improve the public’s health through science and advanced technology. This program is in alignment with the following strategies and objectives of the CDC National Center for Emerging and Zoonotic Infectious Diseases (NCEZID): 1) advance and increase effectiveness of infectious disease laboratory science; 2) develop and apply the science to support and implement new and proven approaches to prevent and control infectious diseases; 3) improve public health laboratory capacity for bioterrorism preparedness and response by developing and validating new tools and tests to aid laboratory detection and identification of new, unknown, emerging, or bioterror disease threats; and 4) support the NCEZID workforce by attracting, maintaining, and developing a highly skilled, motivated and diverse workforce to fulfill the mission of NCEZID.","alternativeNames":["CDC Resident Postdoctoral Program in Microbiology"],"historicalIndex":[{"organizationId":"100007448","actionType":"publish","createdDate":"2017-11-11T05:00:04-05:00","isManual":false,"index":1,"historicalIndexId":"1ea93dcda095b82186dc15b4ad2d9a67","body":"Centers for Disease Control and Prevention - Resident Postdoctoral Program in Microbiology","fiscalYear":2017,"statusCode":null}],"isLatest":true,"programNumber":"93.425","organizationHierarchy":[{"organizationId":"100004222","level":1,"name":"HEALTH AND HUMAN SERVICES, DEPARTMENT OF"},{"organizationId":"100007448","level":2,"name":"CENTERS FOR DISEASE CONTROL AND PREVENTION"}],"postAward":{"audit":{"questions":[{"questionCode":"subpartF","isSelected":"true"},{"questionCode":"subpartD","isSelected":"true"},{"questionCode":"subpartE","isSelected":"true"},{"questionCode":"subpartB","isSelected":"true"},{"questionCode":"subpartC","isSelected":"true"}]},"accomplishments":null},"website":"http://www.cdc.gov","suggestion":{"input":["centers for disease control and prevention - resident postdoctoral program in microbiology","93.425"],"context":{"isActive":true}},"_rScore":0,"preApplication":{"coordination":{"reports":[{"reportCode":"statement","isSelected":false},{"reportCode":"assessment","isSelected":false},{"reportCode":"ExecutiveOrder12372","isSelected":false},{"reportCode":"otherRequired","isSelected":true}],"description":"A Letter of Intent to apply (LOI) may be requested. The LOI is a preliminary, non-binding indication of an organization’s intent to submit an application. The purpose of an LOI is to allow CDC program staff to estimate the number of applicants and plan for the review of submitted applications. If required, instructions for submitting an LOI will be stated in the Notice of Funding Opportunity (NOFO).\r\n"}},"_type":"assistanceListing","indexedDate":"2018-02-27T14:30:33-05:00","relatedPrograms":[{"programNumber":"93.283","id":"e1b05e3b49bc690806682cd8d0f66b69","title":"Centers for Disease Control and Prevention Investigations and Technical Assistance "},{"programNumber":"93.065","id":"3d4c75989cda5fda24a61ef1bbac6ee2","title":"Laboratory Leadership, Workforce Training and Management Development, Improving Public Health Laboratory Infrastructure"},{"programNumber":"93.322","id":"317239dc420f707b64a5b13d950b907d","title":"CSELS Partnership: Strengthening Public Health Laboratories"}],"isFunded":false,"application":{"selectionCriteria":"Applications will be evaluated on the review criteria described in the Notice of Funding Opportunity (NOFO). In general, the review and selection process of complete and responsive applications to the NOFO consists of determination of the technical merit by objective or peer review, and availability of funds. Refer to the NOFO for additional review criteria. The budget is not scored. "},"modifiedDate":"2017-11-11T05:00:03-05:00","_id":"d423160f153b56235b58543b81a6f711","contacts":[{"address":{"zip":"30329","country":"US","streetAddress":"Centers for Disease Control and Prevention\r\n1600 Clifton Road, MS C-18","city":"Atlanta","state":"GA"},"phone":"4046397722","name":"Angela Y. Slaughter","title":"Program Manager"}]},{"postAward":{"audit":{"questions":[{"questionCode":"subpartB","isSelected":"true"},{"questionCode":"subpartD","isSelected":"true"},{"questionCode":"subpartC","isSelected":"true"},{"questionCode":"subpartF","isSelected":"true"},{"questionCode":"subpartE","isSelected":"true"}]},"accomplishments":null},"financial":{"additionalInfo":null,"obligations":[{"assistanceType":{"code":"0003003","value":"Project Grants"},"isRecoveryAct":false,"values":[{"year":2018,"estimate":0},{"actual":0,"flag":"nsi","year":2016},{"year":2017,"estimate":2772731}],"additionalInfo":null}]},"suggestion":{"input":["snap recipient integrity education grant","10.535"],"context":{"isActive":true}},"usage":{"discretionaryFund":"Funds for SNAP Recipient Integrity Education Grant projects are to be used for technical assistance and activities to assist States in developing and improving SNAP recipient fraud prevention education and awareness programs. Applications must comply with criteria listed in the Application Format and Requirements Letter.","loanTerms":null,"restrictions":"Funds for SNAP Recipient Integrity Education Grant projects are to be used for technical assistance and activities to assist States in developing and improving SNAP recipient fraud prevention education and awareness programs. Applications must comply with criteria listed in the Application Format and Requirements Letter.","rules":null},"_rScore":0,"publishDate":"2017-11-11T05:00:02-05:00","_type":"assistanceListing","eligibility":{"beneficiary":{"types":[{"code":"31","value":"Individual/Family"}],"additionalInfo":"Agreements are established between State agencies and FNS. Grants funds can be used for projects educating recipients on eligibility fraud as well as trafficking and misuse of benefits."},"documentation":"Specific instructions provided in Application Format and Requirements Letter.","applicant":{"types":[{"code":"14","value":"State (includes District of Columbia, public institutions of higher education and hospitals)"}],"additionalInfo":"This grant opportunity is open to the 53 State agencies that administer SNAP. FNS will consider only one application per State agency.  There is no State matching requirement for this grant program.  ","assistanceUsageTypes":null}},"assistanceTypes":[{"hierarchy":[{"elementId":"0003","code":"B","level":1,"value":"PROJECT GRANTS"},{"elementId":"0003003","code":"B","level":2,"value":"Project Grants"}]}],"indexedDate":"2018-02-27T14:30:33-05:00","title":"SNAP Recipient Integrity Education Grant","isActive":true,"objective":"The purpose of these grants is to fund technical assistance and activities to assist States in developing and improving SNAP recipient fraud prevention education and awareness programs. Grants funds can be used for projects educating recipients on eligibility fraud as well as trafficking and misuse of benefits.","isFunded":false,"historicalIndex":[{"organizationId":"100076645","actionType":"publish","createdDate":"2017-11-11T05:00:03-05:00","isManual":false,"index":1,"historicalIndexId":"330695dee4bb208d8962c5b4e251ee16","body":"SNAP Recipient Integrity Education Grant","fiscalYear":2017,"statusCode":null}],"application":{"selectionCriteria":"All applications related to the objectives of this grant program and meet the published deadline for submission with all the required documentations will be awarded. "},"isLatest":true,"modifiedDate":"2017-11-11T05:00:02-05:00","programNumber":"10.535","organizationHierarchy":[{"organizationId":"100006809","level":1,"name":"AGRICULTURE, DEPARTMENT OF"},{"organizationId":"100076645","level":2,"name":"FOOD AND NUTRITION SERVICE"}],"_id":"63ac94f6b289c7eca1d801f2c330aa16","contacts":[{"address":{"zip":"22302","country":"US","streetAddress":"3101 Park Center Drive Room 818","city":"Alexandria","state":"VA"},"phone":"7036054385","name":"Jane C. Duffield","title":"Branch Chief, State Admin Branch"}]}]},"_links":{"self":{"href":"https://gsaiae-dev02.reisys.com/sgs/v1/search?index=cfda&index=cfda&q=&q=&page=0&page=0&size=10&sort=-modifiedDate&sort=-modifiedDate&raw=false&is_active=true"}},"page":{"size":10,"totalElements":2251,"totalPages":226,"number":0,"maxAllowedRecords":10000}};
