--- src/patterns/pages/sample/sample.module.ts
@@ -1,32 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-
-import { SamUIKitModule } from '@gsa-sam/sam-ui-elements';
-
-import { SamSampleDemoComponent } from './sample.component';
-import { SamLayoutComponentsModule, DataStore, layoutStore } from '@gsa-sam/sam-ui-elements';
-
-
-@NgModule({
-  imports: [
-    CommonModule,
-    FormsModule,
-    ReactiveFormsModule,
-    SamUIKitModule,
-    SamLayoutComponentsModule
-  ],
-  declarations: [
-    SamSampleDemoComponent
-  ],
-  providers: [
-    {
-      provide: DataStore,
-      useValue: layoutStore
-    }
-  ],
-  exports: [
-    SamSampleDemoComponent
-  ]
-})
-export class SampleDemoModule {}
