+++ src/patterns/pages/workspace/workspace.template.html
@@ -1,0 +1,1 @@
+<div> Workspace working</div>
